Make device-agent run independently of systray
You can now run device-agent as long as you can live without (the interactions of) the systray.
