+++ go.mod
@@ -3,30 +3,40 @@ module github.com/nais/device
 go 1.13
 
 require (
-	cloud.google.com/go v0.65.0
-	cloud.google.com/go/storage v1.10.0
-	github.com/GoogleCloudPlatform/cloudsql-proxy v1.18.0
-	github.com/coreos/go-iptables v0.4.5
+	cloud.google.com/go v0.75.0
+	cloud.google.com/go/storage v1.12.0
+	github.com/GoogleCloudPlatform/cloudsql-proxy v1.19.1
+	github.com/coreos/go-iptables v0.5.0
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible
 	github.com/gen2brain/beeep v0.0.0-20200526185328-e9c15c258e28
-	github.com/getlantern/systray v1.0.3
-	github.com/go-chi/chi v4.0.4+incompatible
+	github.com/getlantern/golog v0.0.0-20201105130739-9586b8bde3a9 // indirect
+	github.com/getlantern/hidden v0.0.0-20201229170000-e66e7f878730 // indirect
+	github.com/getlantern/ops v0.0.0-20200403153110-8476b16edcd6 // indirect
+	github.com/getlantern/systray v1.1.0
+	github.com/go-chi/chi v4.1.2+incompatible
 	github.com/go-chi/jwtauth v4.0.4+incompatible
-	github.com/konsorten/go-windows-terminal-sequences v1.0.2 // indirect
-	github.com/lib/pq v1.8.0
+	github.com/golang/protobuf v1.4.3
+	github.com/gopherjs/gopherjs v0.0.0-20200217142428-fce0ec30dd00 // indirect
+	github.com/lib/pq v1.9.0
 	github.com/nirasan/go-oauth-pkce-code-verifier v0.0.0-20170819232839-0fbfe93532da
-	github.com/prometheus/client_golang v1.6.0
+	github.com/prometheus/client_golang v1.9.0
+	github.com/prometheus/common v0.15.0
+	github.com/prometheus/procfs v0.3.0 // indirect
 	github.com/rjeczalik/notify v0.9.3-0.20181126183243-629144ba06a1
-	github.com/sirupsen/logrus v1.4.2
+	github.com/sirupsen/logrus v1.7.0
 	github.com/spf13/pflag v1.0.5
 	github.com/stretchr/testify v1.5.1
 	github.com/t-tomalak/logrus-easy-formatter v0.0.0-20190827215021-c074f06c5816
 	github.com/titanous/noise v0.0.0-20180327030543-2492fe189ae6
-	golang.org/x/crypto v0.0.0-20200622213623-75b288015ac9
-	golang.org/x/net v0.0.0-20200904194848-62affa334b73
-	golang.org/x/oauth2 v0.0.0-20200902213428-5d25da1a8d43
-	golang.org/x/sys v0.0.0-20200828194041-157a740278f4
-	google.golang.org/api v0.31.0
-	google.golang.org/genproto v0.0.0-20200831141814-d751682dd103
-	google.golang.org/grpc v1.31.1
+	golang.org/x/crypto v0.0.0-20201221181555-eec23a3978ad
+	golang.org/x/mod v0.4.1 // indirect
+	golang.org/x/net v0.0.0-20210119194325-5f4716e94777
+	golang.org/x/oauth2 v0.0.0-20210126194326-f9ce19ea3013
+	golang.org/x/sys v0.0.0-20210124154548-22da62e12c0c
+	golang.org/x/text v0.3.5 // indirect
+	golang.org/x/tools v0.1.0 // indirect
+	google.golang.org/api v0.37.0
+	google.golang.org/genproto v0.0.0-20210126160654-44e461bb6506
+	google.golang.org/grpc v1.35.0
+	google.golang.org/protobuf v1.25.0
 )
